mv more logic to observable package
