+++ library/react-connect-kit/package.json
@@ -10,7 +10,7 @@
   "dependencies": {
     "@w3ux/extension-assets": "^2.2.0",
     "@w3ux/hooks": "^2.1.0",
-    "@w3ux/observables-connect": "^0.9.3",
+    "@w3ux/observables-connect": "^0.9.4",
     "@w3ux/utils": "^2.0.8"
   },
   "devDependencies": {
